+++ cyral/internal/repository/confanalysis/constants.go
@@ -1,0 +1,5 @@
+package confanalysis
+
+const (
+	resourceName = "cyral_repository_conf_analysis"
+)
